+++ src/sjbase/ids.h
@@ -232,7 +232,6 @@
 #define IDO_VIS_STOP            4849
 #define IDO_VIS_OPTIONFIRST     4850    /* range start */
 #define IDO_VIS_OPTIONLAST      4899    /* range end */
-#define IDO_VIS_EMBEDWINDOW     4900
 #define IDO_VIS_HALFSIZE        4901
 #define IDO_VIS_AUTOSWITCHOVER  4902
 #define IDO_VIS_NEXTPRESET      4903
@@ -401,10 +400,7 @@
 
 #define IDMODMSG__VIS_MOD_PRIVATE_1__               11051
 #define IDMODMSG__VIS_MOD_PRIVATE_2__               11052
-#define IDMODMSG__VIS_MOD_PRIVATE_3__               11053
 #define IDMODMSG__VIS_MOD_PRIVATE_4__               11054
-#define IDMODMSG_VIS_BEFORE_REPARENT                11055
-#define IDMODMSG_VIS_AFTER_REPARENT                 11056
 
 #define IDMODMSG_PROGRAM_LOADED                     11060 /* send to indicate, the program is loaded and the main window is shown*/