+++ src/sjmodules/kiosk/monitor_overview.h
@@ -71,11 +71,6 @@ private:
 
 	long            FindMonitorByPoint  (int x, int y) const;
 
-	void            OnMouseRightUp      (wxContextMenuEvent& event) { wxPoint p = ScreenToClient(event.GetPosition()); OnContextMenu(p.x, p.y); }
-	void            OnMouseLeftDClick   (wxMouseEvent& event);
-	void            OnContextMenu       (int x, int y);
-	void            OnMenuSelect        (wxCommandEvent& e);
-
 	long            m_affectedIndex;
 
 	void            DoPaint             (wxDC&);