+++ src/sjmodules/vis/vis_frame.cpp
@@ -46,6 +46,8 @@ SjVisFrame::~SjVisFrame()
 
 BEGIN_EVENT_TABLE(SjVisFrame, wxFrame)
 	//EVT_MENU_RANGE    (IDO_VIS_FIRST__, IDO_VIS_LAST__, SjVisFrame::OnFwdToMainFrame )
+	EVT_ERASE_BACKGROUND(                       SjVisFrame::OnEraseBackground   )
+	EVT_PAINT           (                       SjVisFrame::OnPaint             )
 	EVT_MENU_RANGE      (IDT_FIRST, IDT_LAST,   SjVisFrame::OnFwdToMainFrame    )
 	EVT_CLOSE           (                       SjVisFrame::OnCloseWindow       )
 END_EVENT_TABLE()
@@ -57,11 +59,27 @@ void SjVisFrame::OnFwdToMainFrame(wxCommandEvent& e)
 }
 
 
-void SjVisFrame::OnCloseWindow(wxCloseEvent&)
+void SjVisFrame::OnCloseWindow(wxCloseEvent& e)
 {
 	g_visModule->StopVis();
+	if( e.CanVeto() )
+	{
+		e.Veto();
+	}
 }
 
 
+void SjVisFrame::OnEraseBackground(wxEraseEvent& e)
+{
+}
 
+
+void SjVisFrame::OnPaint(wxPaintEvent& e)
+{
+	wxPaintDC dc(this);
+	wxSize size = GetClientSize();
+	dc.SetBrush(*wxBLACK_BRUSH);
+	dc.SetPen(*wxBLACK_PEN);
+	dc.DrawRectangle(0, 0, size.x, size.y);
+}